Move functionality to mixin
By moving the main functionality of this addon into a Mixin, Components
from other addons can implement these features. For example, other UI
libraries might provide a dropdown Component where this might be a
helpful addition.
