+++ app/mixins/active-link.js
@@ -1,0 +1,3 @@
+import ActiveLinkMixin from 'ember-cli-active-link-wrapper/mixins/active-link';
+
+export default ActiveLinkMixin;
